Fix Unit tests and simplify playback logic (and move it to MainModel)
