+++ CloudMusicPlayer/ViewModels/PlayListInfoModel.swift
@@ -46,39 +46,4 @@ class PlayListInfoModel {
 			}
 		}
 	}
-	
-	func togglePlayer(play: Bool, track: TrackType? = nil) {
-		// if should pause player and track not specified, simply pause and exit
-		if !play { mainModel.player.pause(); return }
-		
-		let player = mainModel.player
-		
-		// check new track specified
-		guard let newTrack = track?.synchronize() else {
-			// if should pause, simply do that
-			if playListActive {
-				// if track not specified and player play current play list, simply resume
-				player.resume(true)
-			} else {
-				// else spart playing current play list
-				mainModel.playPlayList(playList)
-			}
-			return
-		}
-		
-		// if play list not active, simply start playing this play list
-		guard playListActive && player.currentItems.count == playList.items.count else {
-			mainModel.playPlayList(playList, startWith: newTrack); return
-		}
-		
-		if player.current?.streamIdentifier.streamResourceUid == newTrack.uid {
-			// if player current item equals to specified track, resume
-			player.resume(true)
-		} else {
-			// else, set specified track as current and resume
-			if let queueItem = player.getQueueItemByUid(newTrack.uid) {
-				player.play(queueItem)
-			}
-		}
-	}
 }
